Flag operational clarifications to run against local script
Clarified the options to run against your local copy of the update script. As we are working and maintaining it, we often test updates against the local copy of the script we have. Which you may not see us doing this testing, but is happening behind the scenes. 

To stay current it's recommended you update your script regularly. 

These are simply for reference (in case anyone wants to test and propose changes)